Make scripts work on a machine that's not mapit user
These scripts were used to setup a database and machine for a clean
mapit instance running on EC2.

I've now updated the scripts to work against the GDS Dev VMs from a
clean boot. Additionally, as the mapit manage.py isn't on the PATH by
default, this calls it using python.