Allow for detecting the virtualenv directory for mapit
On a deployed machine it is likely to live in /var/apps/mapit/venv as
this is the default path used by our deployment scripts. On the dev VM
it is likely to live in /var/apps/mapit/.venv as that's what our
startup.sh uses (.venv was already in .gitignore for the project). If it's
not in either location you can specify MAPIT_VENV_DIR to tell the script
where to find it.