Download osni from mysociety cache
Now that an OSNI mirror is available on mysociety's cache[1] we can download it and not require the runner of the script to do it manually.  The pacakge is expected to be a zipped tarball that contains metadata.json and a bunch of other zip files that contain the shapefiles from osni.  We also extract the srid for each shapefile from the metadata.json (if present, failing gracefully) so it can be provided to the import script.

The current package was prepared by us and given to mysociety to mirror, it's not clear if future releases will maintain the structure of the metadata.json file, or even the file at all.  Care will need to be taken when using new releases.

[1]: http://parlvid.mysociety.org/os/osni-dec-2015.tar.gz