Move call to show missing codes to end of import
Running `mapit_UK_show_missing_codes` makes the import script fail, because it attempts to fetch the current generation, which has not been activated yet by this point (it happens at the end of the script), so '0' is returned instead of a Generation object.

To fix this we invoke the script to show missing codes at the end of the import (after the current Generation has been created) and we check for all codes, as it serves as a useful summary of the import process.